requirements.api.txt
fastapi
jinja2
itsdangerous
uvicorn[standard]
python-multipart
python-docx
PyMuPDF
pytesseract
Pillow
openai
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
